Complete the JWT & SpringSecurity

+++ src/main/java/com/s1/GESTION_PROFESION/dto/Request/ProductoRequestDTO.java
@@ -1,13 +1,14 @@
 package com.s1.GESTION_PROFESION.dto.Request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.math.BigDecimal;
 
 public record ProductoRequestDTO(
         @Schema(example = "Teclado Mecánico", description = "Lo último en guaracha INC")
         String nombre,
 
         @Schema(example = "85.99", description = "Precio de venta")
-        Double precio,
+        BigDecimal precio,
 
         @Schema(example = "50", description = "Stock")
         Integer stock
