+++ src/main/java/com/s1/GESTION_PROFESION/modelo/Producto.java
@@ -2,15 +2,20 @@ package com.s1.GESTION_PROFESION.modelo;
 
 import jakarta.persistence.*;
 import lombok.*;
+import java.math.BigDecimal;
 
 @Data
 @Entity
 @NoArgsConstructor
 @AllArgsConstructor
 public class Producto {
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
     private String nombre;
-    private Double precio;
+
+    private BigDecimal precio;
+
     private Integer stock;
 }
