+++ src/main/java/com/s1/GESTION_PROFESION/modelo/Venta.java
@@ -1,6 +1,7 @@
 package com.s1.GESTION_PROFESION.modelo;
 
 import jakarta.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.*;
@@ -13,8 +14,11 @@ public class Venta {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
     private LocalDateTime fecha;
-    private Double total;
+
+    private BigDecimal total;
+
     @OneToMany(mappedBy = "venta", cascade = CascadeType.ALL)
     private List<DetalleVenta> detalles;
 }
